create safe utils package for circual-dependency safe utils
