+++ go/controller/internal/utils/safe/README.md
@@ -1,0 +1,2 @@
+# Important
+This package can only use external dependencies. Internal imports must not be used here not to create circular dependencies.
